+++ jetty-util/src/test/java/org/eclipse/jetty/util/statistic/RateStatisticTest.java
@@ -24,7 +24,8 @@ import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.MatcherAssert.assertThat;
+
 
 
 /* ------------------------------------------------------------ */
