+++ tests/test-sessions/test-gcloud-sessions/src/test/java/org/eclipse/jetty/gcloud/session/ClusteredOrphanedSessionTest.java
@@ -22,6 +22,7 @@ package org.eclipse.jetty.gcloud.session;
 import org.eclipse.jetty.server.session.AbstractClusteredOrphanedSessionTest;
 import org.eclipse.jetty.server.session.SessionDataStoreFactory;
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -32,10 +33,25 @@ import org.junit.jupiter.api.Test;
 public class ClusteredOrphanedSessionTest extends AbstractClusteredOrphanedSessionTest
 {
 
+    public static GCloudSessionTestSupport __testSupport;
+
+    @BeforeAll
+    public static void setUp () throws Exception
+    {
+        __testSupport = new GCloudSessionTestSupport();
+        __testSupport.setUp();
+    }
+
+    @AfterAll
+    public static void tearDown () throws Exception
+    {
+        __testSupport.tearDown();
+    }
+
     @AfterAll
     public static void teardown () throws Exception
     {
-        GCloudTestSuite.__testSupport.deleteSessions();
+        __testSupport.deleteSessions();
     }
     
 
@@ -46,7 +62,7 @@ public class ClusteredOrphanedSessionTest extends AbstractClusteredOrphanedSessi
     @Override
     public SessionDataStoreFactory createSessionDataStoreFactory()
     {
-        return GCloudSessionTestSupport.newSessionDataStoreFactory(GCloudTestSuite.__testSupport.getDatastore());
+        return GCloudSessionTestSupport.newSessionDataStoreFactory(__testSupport.getDatastore());
     }
