+++ tests/test-sessions/test-gcloud-sessions/src/test/java/org/eclipse/jetty/gcloud/session/GCloudSessionDataStoreTest.java
@@ -22,7 +22,9 @@ package org.eclipse.jetty.gcloud.session;
 import org.eclipse.jetty.server.session.AbstractSessionDataStoreTest;
 import org.eclipse.jetty.server.session.SessionData;
 import org.eclipse.jetty.server.session.SessionDataStoreFactory;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
 
 /**
  * GCloudSessionDataStoreTest
@@ -32,24 +34,39 @@ import org.junit.jupiter.api.AfterEach;
 public class GCloudSessionDataStoreTest extends AbstractSessionDataStoreTest
 {
 
+    public static GCloudSessionTestSupport __testSupport;
+
+    @BeforeAll
+    public static void setUp () throws Exception
+    {
+        __testSupport = new GCloudSessionTestSupport();
+        __testSupport.setUp();
+    }
+
+    @AfterAll
+    public static void tearDown () throws Exception
+    {
+        __testSupport.tearDown();
+    }
+
     @AfterEach
     public void teardown () throws Exception
     {
-       GCloudTestSuite.__testSupport.deleteSessions();
+       __testSupport.deleteSessions();
     }
     
 
     @Override
     public SessionDataStoreFactory createSessionDataStoreFactory()
     {
-        return GCloudSessionTestSupport.newSessionDataStoreFactory(GCloudTestSuite.__testSupport.getDatastore());
+        return GCloudSessionTestSupport.newSessionDataStoreFactory(__testSupport.getDatastore());
     }
 
   
     @Override
     public void persistSession(SessionData data) throws Exception
     {
-        GCloudTestSuite.__testSupport.createSession(data.getId(), data.getContextPath(), data.getVhost(), data.getLastNode(), data.getCreated(),
+        __testSupport.createSession(data.getId(), data.getContextPath(), data.getVhost(), data.getLastNode(), data.getCreated(),
                                                     data.getAccessed(), data.getLastAccessed(), data.getMaxInactiveMs(), data.getExpiry(), 
                                                     data.getCookieSet(), data.getLastSaved(), data.getAllAttributes());
 
@@ -60,7 +77,7 @@ public class GCloudSessionDataStoreTest extends AbstractSessionDataStoreTest
     public void persistUnreadableSession(SessionData data) throws Exception
     {
 
-        GCloudTestSuite.__testSupport.createSession(data.getId(), data.getContextPath(), data.getVhost(), data.getLastNode(), data.getCreated(),
+        __testSupport.createSession(data.getId(), data.getContextPath(), data.getVhost(), data.getLastNode(), data.getCreated(),
                                                     data.getAccessed(), data.getLastAccessed(), data.getMaxInactiveMs(), data.getExpiry(), 
                                                     data.getCookieSet(), data.getLastSaved(), null);
     }
@@ -69,7 +86,7 @@ public class GCloudSessionDataStoreTest extends AbstractSessionDataStoreTest
     @Override
     public boolean checkSessionExists(SessionData data) throws Exception
     {
-        return GCloudTestSuite.__testSupport.checkSessionExists(data.getId());
+        return __testSupport.checkSessionExists(data.getId());
     }
 
 
@@ -79,7 +96,7 @@ public class GCloudSessionDataStoreTest extends AbstractSessionDataStoreTest
     @Override
     public boolean checkSessionPersisted(SessionData data) throws Exception
     {
-        return GCloudTestSuite.__testSupport.checkSessionPersisted(data);
+        return __testSupport.checkSessionPersisted(data);
     }
 
 }
