+++ tests/test-sessions/test-gcloud-sessions/src/test/java/org/eclipse/jetty/gcloud/session/InvalidationSessionTest.java
@@ -23,6 +23,7 @@ package org.eclipse.jetty.gcloud.session;
 import org.eclipse.jetty.server.session.AbstractClusteredInvalidationSessionTest;
 import org.eclipse.jetty.server.session.SessionDataStoreFactory;
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 
 /**
  * InvalidationSessionTest
@@ -31,11 +32,26 @@ import org.junit.jupiter.api.AfterAll;
  */
 public class InvalidationSessionTest extends AbstractClusteredInvalidationSessionTest
 {
-    
+
+    public static GCloudSessionTestSupport __testSupport;
+
+    @BeforeAll
+    public static void setUp () throws Exception
+    {
+        __testSupport = new GCloudSessionTestSupport();
+        __testSupport.setUp();
+    }
+
+    @AfterAll
+    public static void tearDown () throws Exception
+    {
+        __testSupport.tearDown();
+    }
+
     @AfterAll
     public static void teardown () throws Exception
     {
-        GCloudTestSuite.__testSupport.deleteSessions();
+        __testSupport.deleteSessions();
     }
 
     /** 
@@ -44,6 +60,6 @@ public class InvalidationSessionTest extends AbstractClusteredInvalidationSessio
     @Override
     public SessionDataStoreFactory createSessionDataStoreFactory()
     {
-        return GCloudSessionTestSupport.newSessionDataStoreFactory(GCloudTestSuite.__testSupport.getDatastore());
+        return GCloudSessionTestSupport.newSessionDataStoreFactory(__testSupport.getDatastore());
     }
 }
